requirements.txt
streamlit>=1.32.0
snowflake-connector-python>=3.0.0
pypdf>=3.0.0
